+++ src/core/ip2c/ip2cparser.h
@@ -57,15 +57,7 @@ class IP2CParser : public QObject
 	Q_OBJECT
 
 public:
-	IP2CParser(IP2C *pTargetDatabase);
-
-	/**
-	 * @brief Retrieves the IP2C database this parser operates on.
-	 */
-	IP2C *ip2c() const
-	{
-		return pTargetDatabase;
-	}
+	IP2CParser();
 
 	/**
 	 * @brief For multi-threading purposes. If this is true it is not
@@ -74,6 +66,11 @@ public:
 	bool isParsing() const
 	{
 		return bIsParsing;
+	}
+
+	const QList<IP2C::IP2CData> parsed() const
+	{
+		return parsedValues;
 	}
 
 	bool readDatabase(const QString &filePath);
@@ -88,7 +85,7 @@ signals:
 	 */
 	void parsingFinished(bool bSuccess);
 
-protected:
+private:
 	/**
 	 * @brief Sets states for IP2C when being constructed and destructed.
 	 *
@@ -131,22 +128,14 @@ protected:
 
 	bool bIsParsing;
 	ParsingThread *currentParsingThread;
-
-	/**
-	 * @brief Database to which the IP2C parser will save the data it
-	 * retrieves from IP2C file.
-	 *
-	 * Since IP2CParser is prepared to work in a separate thread it is
-	 * not advised to delete the IP2C object before parsing is complete.
-	 */
-	IP2C *pTargetDatabase;
+	QList<IP2C::IP2CData> parsedValues;
 
 	QMutex thisLock;
 
 	bool doReadDatabase(const QString &filePath);
 	bool readDatabaseVersion2(const QByteArray &dataArray);
 
-protected slots:
+private slots:
 	void parsingThreadFinished();
 };